Modify file and text

See merge request CathayBC/BlockchainDeploymentKit/bdk-quorum!6

+++ src/quorum/command/explorer/delete.ts
@@ -1,11 +1,12 @@
 import { Arguments } from 'yargs'
 import config from '../../config'
 import Explorer from '../../service/explorer'
-import { logger } from '../../../util'
+import { logger, onCancel } from '../../../util'
+import prompts from 'prompts'
 
 export const command = 'delete'
 
-export const desc = 'stop blockchain explorer.'
+export const desc = '刪除現有的 Quorum Explorer.'
 
 export const builder = {}
 
@@ -14,6 +15,19 @@ export const handler = async (argv: Arguments) => {
 
   const explorer = new Explorer(config)
 
-  await explorer.delete()
-  logger.info('Explorer Service delete Successfully')
+  let confirmDelete = true
+
+  const response = await prompts({
+    type: 'confirm',
+    name: 'value',
+    message: 'Confirm to delete Quorum Explorer?',
+    initial: false,
+  }, { onCancel })
+
+  confirmDelete = response.value
+
+  if (confirmDelete) {
+    await explorer.delete()
+    logger.info('Quorum Explorer delete Successfully!')
+  }
 }
